U10416021_HW5_interface-
Using JavaFx to display a bean machine.This program just display the frame and dots,it can't play recently.
BeanMachine contains two classes.
class BeanMachine contains a start method ,which creates a scene,set the title, set the scene and display the stage.
Second class MyPolyLine extend Pane, and it contains a paint method, which create a Polyline, set its fill and stroke.
Then,add the coordinate of eight points to form the polyline. 
And add seven lines to form eight slots.